chapter-5/ticket-generator/requirements.txt
functions-framework==3.3.0
google-cloud-spanner
google-cloud-storage
reportlab
